Bump dependencies and svd files.
This change updates `cortex-m` dependency as well as `svd2rust`, which
required updating xtask source code too. Effectively, `svd2rust` was
the biggest update in terms of dependencies.

This also updates SVD files for all nRF chips to the latest version.

I presume this is going to introduce breaking changes all over the
place, so I also bumped the minor version (which is not exactly
semver, but it felt arrogant to bump the major).